Corregidos errores en la función de hash

+++ Hash/hash.c
@@ -161,8 +161,8 @@ bool buscar_clave(const hash_t* hash, size_t indice_tabla, const char* clave, vo
 size_t hash_djb2(const char *str){
 		unsigned long hash = 5381;
 		int c;
-		while (c = *str++)
-				hash = ((hash << 5) + hash) + c; /* hash * 33 + c */
+		while ((c = *str++))
+				hash = ((hash << 5) + hash) + (size_t)c; /* hash * 33 + c */
 
 		return hash;
 }
